[CELEBORN-1457][0.4] Avoid NPE during shuffle data cleanup
### What changes were proposed in this pull request?

backport https://github.com/apache/celeborn/pull/2553 to `branch-0.4`

Avoid NPE during shuffle data cleanup by checking for null LevelDB.

### Why are the changes needed?
If the LevelDB in StorageManager fails to initialize, the db will be null. This will cause a java.lang.NullPointerException when storageManager.cleanupExpiredShuffleKey(expiredShuffleKeys) is called, and the shuffle data in expiredShuffleKeys will not be cleaned up. The worker's disk may be filled up as a result.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?
Manual Testing

Closes #2565 from cfmcgrady/CELEBORN-1457-branch-0.4.

Authored-by: jiang13021 <jiangyanze.jyz@antgroup.com>
Signed-off-by: Fu Chen <cfmcgrady@gmail.com>