[CELEBORN-1384] Manually excluding workers should not depend on whether the workers are alive
Manually excluding workers should not depend on whether the workers are alive or not for master.

When the workers are offline, master could not add or remove workers through manually excluding workers. Therefore, master should support manually excluding workers no matter whether the workers are alive or not.

No.

Manual test.

Closes #2465 from SteNicholas/CELEBORN-1384.

Authored-by: SteNicholas <programgeek@163.com>
Signed-off-by: Shuang <lvshuang.xjs@alibaba-inc.com>